Update choose-drink.component.ts
The statement !Number.isInteger(this.drink.cups) considered every value of cups, including integral value, as non-integers. I do not know why this is, but to allow the submission of drink orders, I changed that statement to isNaN(this.drink.cups). This checks for null and nonnumeric values, but nonintegral values are now allowed. I am unable to fix this issue at the moment.
Also, an illegitimate value for cups now produces an error message.